+++ src/Database/Esqueleto/Experimental/From/Join.hs
@@ -61,6 +61,18 @@ instance (SqlSelect a ra, SqlSelect b rb) => SqlSelect (a :& b) (ra :& rb) where
         toTuple = const Proxy
     sqlSelectProcessRow = fmap (uncurry (:&)) . sqlSelectProcessRow
 
+-- | Identical to the tuple instance and provided for convenience.
+--
+-- @since 3.5.3.0
+instance (ToAlias a, ToAlias b) => ToAlias (a :& b) where
+    toAlias (a :& b) = (:&) <$> toAlias a <*> toAlias b
+
+-- | Identical to the tuple instance and provided for convenience.
+--
+-- @since 3.5.3.0
+instance (ToAliasReference a, ToAliasReference b) => ToAliasReference (a :& b) where
+    toAliasReference ident (a :& b) = (:&) <$> (toAliasReference ident a) <*> (toAliasReference ident b)
+
 -- | An @ON@ clause that describes how two tables are related. This should be
 -- used as an infix operator after a 'JOIN'. For example,
 --