+++ changelog.md
@@ -5,6 +5,18 @@
         - Deprecate `distinctOn` and `distinctOnOrderBy`. Use the variants
           defined in `PostgreSQL` module instead.
 
+3.5.3.0
+=======
+- @m4dc4p
+  - [#291](https://github.com/bitemyapp/esqueleto/pull/291)
+         - Added `ToAlias` and `ToAliasReference` instaces to the `:&` type, mirroring
+         the tuple instances for the same classes. See [Issue #290](https://github.com/bitemyapp/esqueleto/issues/290)
+         for discussion.       
+- @NikitaRazmakhnin
+  - [#284](https://github.com/bitemyapp/esqueleto/pull/284)
+        - Add PostgreSQL-specific support of VALUES(..) literals
+
+
 3.5.2.2
 =======
 - @NikitaRazmakhnin